renames 'detections' to 'alerts'
