+++ x-pack/plugins/security_solution/cypress/screens/alerts.ts
