+++ x-pack/plugins/security_solution/cypress/screens/alerts_detection_rules.ts
